+++ server/py/services/alerts/main.py
@@ -507,6 +507,6 @@ class Service(framework.service.Service):
 
 
 if __name__ == "__main__":
-    import framework.utils.apiuvicorn as uvicorn
+    import framework.utils.mlrunuvicorn as uvicorn
 
     uvicorn.run(httpdb_config=mlconf.httpdb, service_name="alerts")
